fix: adjust logic offers to card component
